v0.7_01 - Code clarification
ImpactEffect :
    * Removed isTriggerEffect varible for simplifying workflow

* Changed "TrajectoryEffectScript" to "ModularAbility"

ModularAbility :
    * Unsealed and made it abstract
    * Split the projectile portion into ModularProjectileAbility

* Changed namespace "AbilitySystem.TrajectorySystem.ImpactEffect" to "AbilitySystem.TrajectorySystem.ImpactEffects"
