+++ Assets/Scripts/AbilitySystem/Abilities/BombShell/Effects/ModularBomb.asset
@@ -8,8 +8,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fc54b45f841456a45b0bbb411a81e801, type: 3}
-  m_Name: TrajectoryScript
+  m_Script: {fileID: 11500000, guid: fbdb9d56f0b42cd4f88efe4d0ecff978, type: 3}
+  m_Name: ModularBomb
   m_EditorClassIdentifier: 
   effectScriptName: 
   origin: 0
@@ -17,12 +17,12 @@ MonoBehaviour:
   offsetSpace: 1
   direction: {x: 0, y: 0, z: 1}
   directionSpace: 1
-  force: 1000
+  impactEffects:
+  - {fileID: 11400000, guid: 2a876b6346d87304daf5670ec35ea4d2, type: 2}
+  - {fileID: 11400000, guid: a8bb3365906144f42919bed681f5da6e, type: 2}
+  - {fileID: 11400000, guid: 7fbf0c0608d724f4fba7636320b3f132, type: 2}
+  force: 1600
   projectile: {fileID: 11400000, guid: dcbd1ff6e36ea8a4cae27017cf92e65f, type: 2}
   projectileEffects:
   - {fileID: 11400000, guid: aed9e2c497a987845b75294c798355b9, type: 2}
   - {fileID: 11400000, guid: ef47fdc3b23e19441aa129d1f7491984, type: 2}
-  impactEffects:
-  - {fileID: 11400000, guid: a8bb3365906144f42919bed681f5da6e, type: 2}
-  - {fileID: 11400000, guid: 7fbf0c0608d724f4fba7636320b3f132, type: 2}
-  - {fileID: 11400000, guid: 2a876b6346d87304daf5670ec35ea4d2, type: 2}
