--- Assets/Scripts/AbilitySystem/Abilities/BombShell/Effects/TrajectoryScript.asset.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: ffdf673d27a42f04b83925d20ddb8db6
-timeCreated: 1512554013
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
