--- Assets/Scripts/AbilitySystem/Abilities/PeaShooter/Effects/PeaTrajectory.asset.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 0eab60660c23e624caf209b72d454464
-timeCreated: 1511903127
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
