+++ Assets/Scripts/AbilitySystem/Abilities/throwingAxe/Trajectory/ModularAxe.asset.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: 7c0730795ed1e5f46b336a6546f815db
-folderAsset: yes
-timeCreated: 1511867455
+guid: cbc77531964e6f54d9a73ad872da202e
+timeCreated: 1513275679
 licenseType: Free
-DefaultImporter:
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
