--- Assets/Scripts/AbilitySystem/Abilities/throwingAxe/Trajectory/TrajectoryScript.asset.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 05b750e2dd0670448974b302fd109638
-timeCreated: 1512606621
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
