--- Assets/Scripts/AbilitySystem/Resources.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 143cc7066518e7940aefea2ffa77355d
-folderAsset: yes
-timeCreated: 1511287888
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
