--- Assets/Scripts/AbilitySystem/Resources/Affects.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: cb8a8f69a0174c749a6c661f7bd3cdb9
-folderAsset: yes
-timeCreated: 1511894962
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
