--- Assets/Scripts/AbilitySystem/Resources/Affects/New Add Health.asset.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 444738528aaefff41b5cc879ee357512
-timeCreated: 1511894970
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
