--- Assets/Scripts/AbilitySystem/Resources/Affects/New Health Over Time.asset
@@ -1,16 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 36739432d148fca42b71e2032d62ad73, type: 3}
-  m_Name: New Health Over Time
-  m_EditorClassIdentifier: 
-  oneTime: 1
-  lifeTime: 0
-  amount: 0
