--- Assets/Scripts/AbilitySystem/Resources/Effects.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 5b5ff39b26bcba3439086c4f76888e9c
-folderAsset: yes
-timeCreated: 1511288038
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
