--- Assets/Scripts/AbilitySystem/Resources/Effects/ModularTrajectoryEffect.asset
@@ -1,20 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5a7cdcbc554f2034b9a123049d7fa463, type: 3}
-  m_Name: ModularTrajectoryEffect
-  m_EditorClassIdentifier: 
-  effectName: 
-  effectDescription: 
-  maxStacks: 1
-  stackTimer: 5
-  stackCommitment: 2
-  effectScripts:
-  - {fileID: 11400000, guid: b11c167475e5d5e478d88a985bf82ca3, type: 2}
