--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: d3964a857867e66499bf89196ee26ffa
-folderAsset: yes
-timeCreated: 1511727040
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
