--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/CollisionChecks/FilterByObjectMetaData.asset.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: a62999679b76c074eaecd598dc8993ba
-timeCreated: 1511883347
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
