--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/CollisionChecks/FilterByTag.asset
@@ -1,16 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: dab816bfb8f36e14f984326f4a8754dd, type: 3}
-  m_Name: FilterByTag
-  m_EditorClassIdentifier: 
-  tagsToFilter:
-  - Ground
-  exclude: 0
