--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/CollisionChecks/FilterByTag.asset.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 613a884591140114e9258a70bf5ccbc0
-timeCreated: 1511882918
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
