--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/ExampleTrajectoryEffectScript.asset
@@ -1,26 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fc54b45f841456a45b0bbb411a81e801, type: 3}
-  m_Name: ExampleTrajectoryEffectScript
-  m_EditorClassIdentifier: 
-  effectScriptName: 
-  trajectoryType: 1
-  origin: 0
-  originOffset: {x: 0, y: 0, z: 6}
-  offsetSpace: 1
-  direction: {x: 0, y: 0, z: 1}
-  directionSpace: 1
-  force: 500
-  projectile: {fileID: 11400000, guid: af36a9a85d5932d47914b67f8609d687, type: 2}
-  projectileEffects:
-  - {fileID: 11400000, guid: 6b7b4d86cc6abdb4f89dd000bfe90d5f, type: 2}
-  collisionChecks:
-  - {fileID: 11400000, guid: a62999679b76c074eaecd598dc8993ba, type: 2}
