--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/ExampleTrajectoryEffectScript.asset.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: b11c167475e5d5e478d88a985bf82ca3
-timeCreated: 1511727019
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
