--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/ProjectileEffects.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 6eabbf247ada1f44b90972bbeea18017
-folderAsset: yes
-timeCreated: 1511857123
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
