--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/ProjectileEffects/ExampleProjectileEffect.asset
@@ -1,17 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7726bbe3ffda48d4f9fe863c59f4745d, type: 3}
-  m_Name: ExampleProjectileEffect
-  m_EditorClassIdentifier: 
-  toInitializeEffects:
-  - {fileID: 100004, guid: ed30ba2ebe7b1ff4cbc34e610f615eea, type: 3}
-  toFollowEffects:
-  - {fileID: 100004, guid: ed30ba2ebe7b1ff4cbc34e610f615eea, type: 3}
