--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/ProjectileEffects/ExampleProjectileEffect.asset.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 6b7b4d86cc6abdb4f89dd000bfe90d5f
-timeCreated: 1511857260
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
