--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/Projectiles.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 039c6b99f37fac74a8175cf16e81b451
-folderAsset: yes
-timeCreated: 1511781356
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
