--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/Projectiles/ImpactEffects.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: ace59b4d1101b2840a002df16009ff57
-folderAsset: yes
-timeCreated: 1511894842
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
