--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/Projectiles/ImpactEffects/New Give Object Affect.asset
@@ -1,18 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 633f232377ce1cd40b8b04920435ace7, type: 3}
-  m_Name: New Give Object Affect
-  m_EditorClassIdentifier: 
-  isTriggerEffect: 0
-  affectsToSender:
-  - {fileID: 11400000, guid: 444738528aaefff41b5cc879ee357512, type: 2}
-  affectsToHit:
-  - {fileID: 11400000, guid: 12a847ec622cec546af8ee2d3db03602, type: 2}
