--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/Projectiles/ImpactEffects/New Give Object Affect.asset.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 2c50eff4febc8d646945b962b4233b56
-timeCreated: 1511894904
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
