--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/Projectiles/ImpactEffects/ToOther.asset
@@ -1,17 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 17ca6fca757a83c4e990dcf4e8267a3a, type: 3}
-  m_Name: ToOther
-  m_EditorClassIdentifier: 
-  oneTime: 1
-  lifeTime: 0
-  doneTime: 0
-  amount: 5
