--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/Projectiles/ImpactEffects/ToOther.asset.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 12a847ec622cec546af8ee2d3db03602
-timeCreated: 1511899366
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
