--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/Projectiles/New Projectile.asset
@@ -1,18 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6565131259fcf5c47a4f39b086ff883f, type: 3}
-  m_Name: New Projectile
-  m_EditorClassIdentifier: 
-  projectilePrefab: {fileID: 1295420332119136, guid: e86ffe03747b8094ea7ea98947dcb6f3,
-    type: 2}
-  impactEffects:
-  - {fileID: 11400000, guid: 2c50eff4febc8d646945b962b4233b56, type: 2}
-  instantiatedObject: {fileID: 0}
