--- Assets/Scripts/AbilitySystem/Resources/Effects/Trajectory Effect Scripts/Projectiles/New Projectile.asset.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: af36a9a85d5932d47914b67f8609d687
-timeCreated: 1511886106
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
