+++ Assets/Scripts/AbilitySystem/Scripts/TrajectoryEffectScript/Core/Assets/ModularProjectileAbility.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: fbdb9d56f0b42cd4f88efe4d0ecff978
+timeCreated: 1513273663
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
